Move blocks into blocks file and read more link to landing + minor changes in wording/naming
